separated out input controller
